+++ packages/flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/flp-config-sub-generator
 
+## 0.1.67
+
+### Patch Changes
+
+-   Updated dependencies [e32dd29]
+    -   @sap-ux/app-config-writer@0.5.51
+
+## 0.1.66
+
+### Patch Changes
+
+-   @sap-ux/flp-config-inquirer@0.2.83
+-   @sap-ux/app-config-writer@0.5.50
+
 ## 0.1.65
 
 ### Patch Changes